requirements.txt
jsonschema
pillow
pyyaml
svgpathtools
